moved util into same dir as main package
