+++ src/codeoceanauxtools/get_cpu_count.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
-import aux_tools_utils.util
+import util
 
 
 def main():
-    return aux_tools_utils.util.get_cpu_limit()
+    return util.get_cpu_limit()
 
 
 if __name__ == "__main__":
