+++ src/codeoceanauxtools/get_dir_contents.py
@@ -2,18 +2,18 @@
 import sys
 
 from get_logger import LOGGER
-import aux_tools_utils.util
+import util
 
 
 def main(argv=None):
     if len(argv) > 1:
         return (
             f"*** These are the current files in the {argv[1]} directory\n"
-            + f"{aux_tools_utils.util.get_dir_contents(argv[1])}"
+            + f"{util.get_dir_contents(argv[1])}"
         )
     return (
         "*** These are the current files in the ../data directory\n"
-        + f"{aux_tools_utils.util.get_dir_contents()}"
+        + f"{util.get_dir_contents()}"
     )
