+++ src/codeoceanauxtools/get_groups.py
@@ -2,13 +2,13 @@
 import sys
 
 from get_logger import LOGGER
-import aux_tools_utils.util
+import util
 
 
 def main(argv=None):
     if len(argv) > 1:
-        return aux_tools_utils.util.get_groups(argv[1])
-    return aux_tools_utils.util.get_groups()
+        return util.get_groups(argv[1])
+    return util.get_groups()
 
 
 if __name__ == "__main__":
